Fixed some syntaxing and created a global file that holds repeated funtions
